requirements.txt
pandas==2.2.2
python_dateutil==2.9.0.post0
Requests==2.31.0
tkmacosx==1.0.5
websocket_client==1.7.0
